refactor: change data access method to local save
Change the data access of user, post to save by sqlite. By the way, the user is much modified, post is not yet. The register can perform but it seem like error as new user can not have exactly role. Change the ddatatype of avata from user to binary for many use.